move buildAnsibleJobStatus back to ansibleJobHIstoryDef
